Fixes #99: Create separate Android module for tor installer
