+++ android/build.gradle
@@ -70,9 +70,10 @@ dependencies {
     implementation 'net.freehaven.tor.control:jtorctl:0.2'
     implementation 'org.slf4j:slf4j-api:1.7.25'
     implementation 'org.slf4j:slf4j-android:1.7.25'
-    implementation 'org.torproject:tor-android-binary:0.3.5.8-rc-v2'
     implementation 'com.android.support:appcompat-v7:26.1.0'
     testImplementation 'junit:junit:4.12'
+    androidTestImplementation project(':android_tor_installer')
+    androidTestImplementation 'org.torproject:tor-android-binary:0.3.5.8-rc-v2'
     androidTestImplementation 'com.android.support.test:runner:1.0.1'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.1'
 }
@@ -83,7 +84,6 @@ repositories {
     google()
     jcenter()
     maven { url "https://raw.githubusercontent.com/guardianproject/gpmaven/master" }
-
 }
 
 /**
