+++ android_tor_installer/src/main/java/com/msopentech/thali/android/installer/AndroidTorInstaller.java
@@ -10,33 +10,41 @@ MERCHANTABLITY OR NON-INFRINGEMENT.
 
 See the Apache 2 License for the specific language governing permissions and limitations under the License.
 */
-package com.msopentech.thali.android.toronionproxy;
+package com.msopentech.thali.android.installer;
 
 
 import android.content.Context;
 import android.util.Log;
 import com.msopentech.thali.toronionproxy.TorInstaller;
-import com.msopentech.thali.toronionproxy.android.R;
 import org.torproject.android.binary.TorResourceInstaller;
 
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.io.InputStream;
 import java.util.concurrent.TimeoutException;
 
 /**
  * Installs Tor for an Android app. This is a wrapper around the <code>TorResourceInstaller</code>.
+ *
+ * Since this class only deals with installing Tor, it is up to the developer to implement
+ * the <code>openBridgesStream</code> which will give the bridges for pluggable transports. A
+ * typical implementation looks like:
+ *
+ * <code>
+ *     public InputStream openBridgesStream() throws IOException {
+ *         return context.getResources().openRawResource(R.raw.bridges);
+ *     }
+ * </code>
  */
-public class AndroidTorInstaller extends TorInstaller {
+public abstract class AndroidTorInstaller extends TorInstaller {
 
     private final TorResourceInstaller resourceInstaller;
 
     private static final String TAG = "TorInstaller";
 
-    private final Context context;
+    protected final Context context;
 
-    private File torrcFile;
+    protected File torrcFile;
 
    /**
      * The configDir will be the location of tor configuration files. It contains the files, geoip, geoip6,
@@ -79,10 +87,5 @@ public class AndroidTorInstaller extends TorInstaller {
             Log.w(TAG, "Failed to setup tor: " + e.getMessage());
             throw new IOException(e);
         }
-    }
-
-    @Override
-    public InputStream openBridgesStream() throws IOException {
-        return context.getResources().openRawResource(R.raw.bridges);
     }
 }
