+++ android_tor_installer/src/main/res/values/strings.xml
@@ -1,0 +1,3 @@
+<resources>
+    <string name="app_name">Android Tor Installer</string>
+</resources>
